requirements.txt
cryptg==0.4.0
Pillow==10.1.0
pyaes==1.6.1
pyasn1==0.5.0
python-dotenv==1.0.0
rsa==4.9
Telethon==1.32.0
tzdata==2023.3
